LayoutTests/platform/glib/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-inline-editing-host.tentative_white-space=pre-wrap&display=list-item&method=enter-expected.txt

abcd

FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line breaker should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> (defaultParagraphSeparator=div) value "<span contenteditable=\"\"></span>" not in array ["<span contenteditable=\"\">\n<br></span>", "<span contenteditable=\"\">\n\n</span>"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> followed by a <br> (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> followed by a <br> (defaultParagraphSeparator=div) value "<span contenteditable=\"\"></span><br>" not in array ["<span contenteditable=\"\">\n\n</span><br>", "<span contenteditable=\"\">\n<br></span><br>"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> followed by text (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> followed by text (defaultParagraphSeparator=div) value "<span contenteditable=\"\"></span>abc" not in array ["<span contenteditable=\"\">\n\n</span>abc", "<span contenteditable=\"\">\n<br></span>abc"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}<br></span> (defaultParagraphSeparator=div) assert_equals: A <br> should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}<br></span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">\n<br></span>" but got "<span contenteditable=\"\"><br></span>"
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">[]abcd</span> (defaultParagraphSeparator=div) assert_equals: A linefeed should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">[]abcd</span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">\nabcd</span>" but got "<span contenteditable=\"\">abcd</span>"
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">ab[]cd</span> (defaultParagraphSeparator=div) assert_equals: A linefeed should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">ab[]cd</span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">ab\ncd</span>" but got "<span contenteditable=\"\">abcd</span>"
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> (defaultParagraphSeparator=div) value "<span contenteditable=\"\">abcd</span>" not in array ["<span contenteditable=\"\">abcd\n<br></span>", "<span contenteditable=\"\">abcd\n\n</span>"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=div) value "<span contenteditable=\"\">abcd</span><br>" not in array ["<span contenteditable=\"\">abcd\n<br></span><br>", "<span contenteditable=\"\">abcd\n\n</span><br>"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> followed by text (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> followed by text (defaultParagraphSeparator=div) value "<span contenteditable=\"\">abcd</span>efg" not in array ["<span contenteditable=\"\">abcd\n<br></span>efg", "<span contenteditable=\"\">abcd\n\n</span>efg"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line breaker should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> (defaultParagraphSeparator=p) value "<span contenteditable=\"\"></span>" not in array ["<span contenteditable=\"\">\n<br></span>", "<span contenteditable=\"\">\n\n</span>"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> followed by a <br> (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> followed by a <br> (defaultParagraphSeparator=p) value "<span contenteditable=\"\"></span><br>" not in array ["<span contenteditable=\"\">\n\n</span><br>", "<span contenteditable=\"\">\n<br></span><br>"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> followed by text (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}</span> followed by text (defaultParagraphSeparator=p) value "<span contenteditable=\"\"></span>abc" not in array ["<span contenteditable=\"\">\n\n</span>abc", "<span contenteditable=\"\">\n<br></span>abc"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}<br></span> (defaultParagraphSeparator=p) assert_equals: A <br> should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">{}<br></span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">\n<br></span>" but got "<span contenteditable=\"\"><br></span>"
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">[]abcd</span> (defaultParagraphSeparator=p) assert_equals: A linefeed should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">[]abcd</span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">\nabcd</span>" but got "<span contenteditable=\"\">abcd</span>"
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">ab[]cd</span> (defaultParagraphSeparator=p) assert_equals: A linefeed should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">ab[]cd</span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">ab\ncd</span>" but got "<span contenteditable=\"\">abcd</span>"
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> (defaultParagraphSeparator=p) value "<span contenteditable=\"\">abcd</span>" not in array ["<span contenteditable=\"\">abcd\n<br></span>", "<span contenteditable=\"\">abcd\n\n</span>"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=p) value "<span contenteditable=\"\">abcd</span><br>" not in array ["<span contenteditable=\"\">abcd\n<br></span><br>", "<span contenteditable=\"\">abcd\n\n</span><br>"]
FAIL insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> followed by text (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertParagraph in <span contenteditable style="display:list-item;white-space:pre-wrap">abcd[]</span> followed by text (defaultParagraphSeparator=p) value "<span contenteditable=\"\">abcd</span>efg" not in array ["<span contenteditable=\"\">abcd\n<br></span>efg", "<span contenteditable=\"\">abcd\n\n</span>efg"]
